requirements.txt
flask==2.0.1
google-api-python-client==2.108.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
pandas==1.3.5
nltk==3.6.7
python-dotenv==0.19.2
python-dateutil==2.8.2
requests==2.26.0
beautifulsoup4==4.10.0
scikit-learn==1.0.2
textblob==0.17.1
email-validator==1.1.3
openpyxl==3.0.10
